+++ lib/constants/constants.dart
@@ -1,3 +1,6 @@
+// TODO: Remember to update here
+const appVersion = '0.10.5';
+
 const allowedNoteExtensions = ['.md', '.markdown', '.txt', '.me'];
 const allowedImageExtensions = ['.jpg', '.jpeg', '.png', '.bmp', '.gif'];
 const allowedPdfExtensions = ['.pdf'];