+++ lib/ui/screens/about/about_screen.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:package_info_plus/package_info_plus.dart';
 
 import 'package:printnotes/ui/components/app_bar_drag_wrapper.dart';
 import 'package:printnotes/ui/components/centered_page_wrapper.dart';
@@ -9,26 +8,10 @@ import 'package:printnotes/ui/widgets/menu_tile.dart';
 
 import 'package:printnotes/utils/handlers/open_url_link.dart';
 
-class AboutScreen extends StatefulWidget {
+import 'package:printnotes/constants/constants.dart';
+
+class AboutScreen extends StatelessWidget {
   const AboutScreen({super.key});
-
-  @override
-  State<AboutScreen> createState() => _AboutScreenState();
-}
-
-class _AboutScreenState extends State<AboutScreen> {
-  String appVersion = '0.0.0';
-
-  @override
-  void initState() {
-    _getPackageInfo();
-    super.initState();
-  }
-
-  void _getPackageInfo() async {
-    final package = await PackageInfo.fromPlatform();
-    setState(() => appVersion = package.version);
-  }
 
   @override
   Widget build(BuildContext context) {