+++ lib/utils/storage_system.dart
@@ -2,7 +2,6 @@ import 'dart:io';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:path/path.dart' as path;
-// import 'package:docman/docman.dart';
 import 'package:provider/provider.dart';
 
 import 'package:printnotes/providers/settings_provider.dart';
@@ -134,7 +133,7 @@ class StorageSystem {
 
       String contents = await file.readAsString();
 
-      final matches = RegExp(r'#\w+').allMatches(contents);
+      final matches = RegExp(r'^\s*(#\w+)').allMatches(contents);
 
       for (final match in matches) {
         final tag = contents.substring(match.start, match.end);